Check scan plugin is available before suggesting to use it
